(refactor): moved graphql plugins inside the 'internal' package
