+++ amf-graphql/shared/src/main/scala/amf/graphql/internal/spec/plugins/parse/GraphQLParsePlugin.scala
@@ -1,4 +1,4 @@
-package amf.graphqlfederation.plugins.parse
+package amf.graphql.internal.spec.plugins.parse
 
 import amf.antlr.client.scala.parse.document.AntlrParsedDocument
 import amf.apicontract.internal.plugins.ApiParsePlugin
@@ -13,12 +13,13 @@ import amf.core.client.scala.parse.document.{
 }
 import amf.core.internal.parser.Root
 import amf.core.internal.remote.{GraphQL, Spec, Syntax}
+import amf.graphql.internal.spec.context.GraphQLWebApiContext
 import amf.graphql.internal.spec.document.GraphQLBaseDocumentParser
 import amf.graphql.internal.spec.parser.syntax.GraphQLASTParserHelper
-import amf.graphql.internal.spec.parser.syntax.TokenTypes.{DEFINITION, TYPE_SYSTEM_DEFINITION}
-import amf.graphqlfederation.internal.spec.context.GraphQLFederationWebApiContext
+import amf.graphql.internal.spec.parser.syntax.TokenTypes.{DEFINITION, DOCUMENT, TYPE_SYSTEM_DEFINITION}
+import org.mulesoft.antlrast.ast.Node
 
-object GraphQLFederationParsePlugin extends ApiParsePlugin with GraphQLASTParserHelper {
+object GraphQLParsePlugin extends ApiParsePlugin with GraphQLASTParserHelper {
   override def spec: Spec = GraphQL
 
   override def parse(document: Root, ctx: ParserContext): BaseUnit = {
@@ -28,8 +29,8 @@ object GraphQLFederationParsePlugin extends ApiParsePlugin with GraphQLASTParser
   override def referenceHandler(eh: AMFErrorHandler): ReferenceHandler =
     (_: ParsedDocument, _: ParserContext) => CompilerReferenceCollector()
 
-  private def context(document: Root, ctx: ParserContext): GraphQLFederationWebApiContext = {
-    new GraphQLFederationWebApiContext(
+  private def context(document: Root, ctx: ParserContext): GraphQLWebApiContext = {
+    new GraphQLWebApiContext(
       ctx.rootContextDocument,
       ctx.refs,
       ctx.parsingOptions,
@@ -51,7 +52,9 @@ object GraphQLFederationParsePlugin extends ApiParsePlugin with GraphQLASTParser
     }
   }
 
-  private def isGraphQL(doc: AntlrParsedDocument): Boolean = collect(doc.ast.root(), Seq(DEFINITION)).nonEmpty
+  private def isGraphQL(doc: AntlrParsedDocument): Boolean = {
+    collect(doc.ast.root(), Seq(DEFINITION, TYPE_SYSTEM_DEFINITION)).exists(e => e.name == TYPE_SYSTEM_DEFINITION)
+  }
 
   override def withIdAdoption: Boolean = true // TODO pending analysis and parsing cleanup
 }
